+++ plugins/actian-design-system/tests/renderers/appearance-variant-realdata.test.js
@@ -27,7 +27,7 @@ var test = require("node:test");
 var assert = require("node:assert/strict");
 var fs = require("fs");
 var path = require("path");
-var ds = require("../../scripts/renderers/html-renderers/ds-html-map.js");
+var ds = require("../../scripts/lib/renderer.js").dsHtmlMap;
 
 var ANATOMY_DIR = path.join(__dirname, "../../vendor/components/dist/anatomy");